integrated tested code into new frames
