+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 import {BrowserRouter as Router, Routes, Route} from 'react-router-dom';
 import Home from './Pages/Front.js';
-import ServerDeploy from './Pages/ServerDeploy';
+import ServerDashPage from './Pages/ServerDashPage';
 import Navbar from './Components/Navbar';
 
 function App() {
@@ -12,7 +12,7 @@ function App() {
       <Navbar />
       <Routes>
           <Route path='/' element={<Home />} />
-          <Route path='/ServerDeploy' element={<ServerDeploy />} />
+          <Route path='/ServerDeploy' element={<ServerDashPage />} />
         </Routes>
      </Router>
   );
