--- src/Components/ServerDeploy/serverInfo.js
@@ -1,19 +1,0 @@
-import React, { Component } from 'react'
-
-import { FormH1,  Form, EmailRow, ButtonLink} from './serverElements.js';
-
-
- const Status = () => {
-    
-
-    {
-        return (<>
-                    <Form >
-                            <FormH1>Server Status</FormH1>
-                    </Form>
-                </>
-            )
-    }
-}
-
-export default Status;
