+++ src/Pages/Front.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import Navbar from '../Components/Navbar';
 import HomePage from '../Components/MainSection';
 const Home = () => {
