+++ src/Pages/ServerDashPage.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import ServerDashboard from '../Components/ServerDashboard'
+
+const ServerDashPage = () => {
+  return (
+  <>
+    <ServerDashboard />
+  </>
+  );
+};
+
+export default ServerDashPage;
