--- src/Pages/ServerDeploy.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import ServerStation from '../Components/ServerDeploy'
-
-const ServerDeploy = () => {
-  return (
-  <>
-    <ServerStation />
-  </>
-  );
-};
-
-export default ServerDeploy;
